add kkm (fix subject option desc and judul)

+++ database/migrations/2025_04_15_080127_create_report_template_kelas_table.php
@@ -8,15 +8,17 @@ class CreateReportTemplateKelasTable extends Migration
 {
     public function up()
     {
-        Schema::create('report_template_kelas', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('report_template_id')->constrained()->onDelete('cascade');
-            $table->foreignId('kelas_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-            
-            // Tambahkan indeks untuk performa
-            $table->index(['report_template_id', 'kelas_id']);
-        });
+        if (!Schema::hasTable('report_template_kelas')) {
+            Schema::create('report_template_kelas', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('report_template_id')->constrained()->onDelete('cascade');
+                $table->foreignId('kelas_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+                
+                // Tambahkan indeks untuk performa
+                $table->index(['report_template_id', 'kelas_id']);
+            });
+        }
     }
 
     public function down()
